\name{probability-package}
\alias{probability-package}
\alias{probability}
\docType{package}
\title{
\packageTitle{probability}
}
\description{
\packageDescription{probability}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{probability}
\packageIndices{probability}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{probability}

Maintainer: \packageMaintainer{probability}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# simple examples of the most important functions
}
